# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THB Nexus AI Trader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface RegisterResult {
  message: string;
  email: string;
}

export interface OtpInput {
  email: string;
  /**
     * @minLength 6
     * @maxLength 6
     */
  otp: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  email: string;
  name: string;
  isAdmin: boolean;
  isVerified: boolean;
  createdAt: string;
}

export interface AuthResult {
  user: User;
  message: string;
}

export interface Ticker {
  symbol: string;
  price: number;
  change: number;
  changePercent: number;
  high24h: number;
  low24h: number;
  volume: number;
  direction: string;
}

export interface MarketSummary {
  totalVolume: number;
  bullishCount: number;
  bearishCount: number;
  lastUpdated: string;
}

export interface BotStartInput {
  strategy: string;
  riskLevel: string;
  symbols?: string[];
}

export interface BotStatus {
  isRunning: boolean;
  /** @nullable */
  strategy: string | null;
  /** @nullable */
  riskLevel: string | null;
  symbols: string[];
  tradesExecuted: number;
  /** @nullable */
  startedAt: string | null;
  profitLoss: number;
  status: string;
}

export interface Mt5Input {
  brokerName: string;
  loginId: string;
  password: string;
  server: string;
}

export interface Mt5Connection {
  isConnected: boolean;
  /** @nullable */
  brokerName: string | null;
  /** @nullable */
  loginId: string | null;
  /** @nullable */
  server: string | null;
  bridgeRunning: boolean;
  /** @nullable */
  connectedAt: string | null;
  status: string;
}

export interface DepositInput {
  /** @minimum 1 */
  amount: number;
  /** @nullable */
  txHash?: string | null;
}

export interface Deposit {
  id: string;
  userId: string;
  amount: number;
  currency: string;
  status: string;
  /** @nullable */
  txHash: string | null;
  createdAt: string;
}

export interface Balance {
  usdt: number;
  totalDeposited: number;
  totalPnl: number;
}

export interface Trade {
  id: string;
  userId: string;
  symbol: string;
  direction: string;
  entryPrice: number;
  /** @nullable */
  exitPrice: number | null;
  lots: number;
  /** @nullable */
  profitLoss: number | null;
  status: string;
  openedAt: string;
  /** @nullable */
  closedAt: string | null;
}

export interface TradeStats {
  totalTrades: number;
  winningTrades: number;
  losingTrades: number;
  winRate: number;
  totalPnl: number;
  bestTrade: number;
  worstTrade: number;
  avgProfit: number;
}