# lib/api-zod/src/generated/types/balance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THB Nexus AI Trader API
 * OpenAPI spec version: 0.1.0
 */

export interface Balance {
  usdt: number;
  totalDeposited: number;
  totalPnl: number;
}