# lib/api-zod/src/generated/types/botStartInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THB Nexus AI Trader API
 * OpenAPI spec version: 0.1.0
 */

export interface BotStartInput {
  strategy: string;
  riskLevel: string;
  symbols?: string[];
}