# lib/api-zod/src/generated/types/mt5Connection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THB Nexus AI Trader API
 * OpenAPI spec version: 0.1.0
 */

export interface Mt5Connection {
  isConnected: boolean;
  /** @nullable */
  brokerName: string | null;
  /** @nullable */
  loginId: string | null;
  /** @nullable */
  server: string | null;
  bridgeRunning: boolean;
  /** @nullable */
  connectedAt: string | null;
  status: string;
}