# lib/api-zod/src/generated/types/tradeStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THB Nexus AI Trader API
 * OpenAPI spec version: 0.1.0
 */

export interface TradeStats {
  totalTrades: number;
  winningTrades: number;
  losingTrades: number;
  winRate: number;
  totalPnl: number;
  bestTrade: number;
  worstTrade: number;
  avgProfit: number;
}